generative_ai/compliance-agent-mvp-main/server/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cobble==0.1.4
colorama==0.4.6
datarobot==3.9.1
distro==1.9.0
fastapi==0.119.0
griffe==1.14.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
jiter==0.11.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mammoth==1.7.1
markdownify==0.14.1
mcp==1.17.0
numpy==2.3.4
openpyxl==3.1.2
openai==1.109.1
openai-agents==0.3.3
pandas==2.3.3
pydantic==2.12.2
pydantic-settings==2.11.0
pydantic_core==2.41.4
pypdf==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sse-starlette==3.0.2
starlette==0.48.0
StrEnum==0.4.15
tqdm==4.67.1
trafaret==2.1.1
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
xlrd==2.0.1
